requirements.txt
discord.py==1.5.1
requests==2.20.0
pandas==1.1.1
pytz==2020.1
disputils==0.1.3
tabulate==0.8.7
boto3==1.16.56
beautifulsoup4==4.9.3
discord==1.0.1
python-decouple==3.4
